Notify connection disconnect during error handling of new connection from a worker when there is already a known (but actually disconnected) connection to the same worker.
Additionally, match pb.loseConnection with null.loseConnection() by calling self.notifyDisconnected() to make sure the connection is shut down properly.

This function is called both places to make sure the old connection is actually shut down.

This would otherwise cause a case of worker "knowing" it was "connected", while
the manager equally "know" that the worker was "offline" because it was not
attached to any builders, all the while the low-level connection maintenance
functionality kept sending and receiving keep-alives, thus keeping the unattached
connection open. (#4678)